* 0.05 releng ; make the tests not kill shells for people without trac-admin
git-svn-id: svn://svn.bestpractical.com/Net-Trac/trunk@17897 e417ac7c-1bcc-0310-8ffa-8f5827389a85

+++ t/setup_trac.pl
@@ -102,6 +102,7 @@ close $htpasswd;
 
 sub kill_trac {
     my $self = shift;
+    return unless $self->pid;
     kill 1, $self->pid;
 
 }
